server/requirements.txt
Flask==2.1.2
Flask-MQTT==1.2.1
Flask-JWT-Extended==4.3.1
Flask-PyMongo==2.3.0
Flask-Cors==3.0.10
PyMongo==3.12.1
dnspython>=2.1.0,<3.0
mongoengine==0.24.2
requests==2.26.0
pydot==1.4.2
python-dotenv==0.19.2
Werkzeug<3.0
pytest==8.3.3
mongomock==4.1.2